+++ framework-utilities/test-utils/test-utils-core/src/main/java/uk/gov/justice/services/test/utils/core/messaging/JmsSessionFactory.java
@@ -1,34 +1,48 @@
 package uk.gov.justice.services.test.utils.core.messaging;
 
+import static java.lang.String.format;
+import static javax.jms.Session.AUTO_ACKNOWLEDGE;
+
 import javax.jms.Connection;
 import javax.jms.JMSException;
 import javax.jms.Session;
 
 import org.apache.activemq.artemis.jms.client.ActiveMQConnectionFactory;
 
-public class JmsSessionFactory {
+public class JmsSessionFactory implements AutoCloseable {
+
+    private final ActiveMQConnectionFactory activeMQConnectionFactory;
 
     private Session session;
-    private ActiveMQConnectionFactory connectionFactory;
     private Connection connection;
 
-    public Session session(String queueUri) {
-        connectionFactory = new ActiveMQConnectionFactory(queueUri);
+    public JmsSessionFactory() {
+        this(new ActiveMQConnectionFactory());
+
+    }
+    
+    public JmsSessionFactory(final ActiveMQConnectionFactory activeMQConnectionFactory) {
+        this.activeMQConnectionFactory = activeMQConnectionFactory;
+    }
+
+    public Session session(final String queueUri) {
 
         try {
-            connection = connectionFactory.createConnection();
+            activeMQConnectionFactory.setBrokerURL(queueUri);
+            connection = activeMQConnectionFactory.createConnection();
             connection.start();
-            session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
+            session = connection.createSession(false, AUTO_ACKNOWLEDGE);
             return session;
-        } catch (JMSException e) {
-            throw new MessageConsumerException("Unable to create JMS session", e);
+        } catch (final JMSException e) {
+            throw new MessageConsumerException(format("Failed to create JMS session for queue uri '%s'", queueUri), e);
         }
     }
 
+    @Override
     public void close() {
         doClose(session);
         doClose(connection);
-        doClose(connectionFactory);
+        doClose(activeMQConnectionFactory);
     }
 
     private void doClose(final AutoCloseable closeable) {
@@ -38,8 +52,6 @@ public class JmsSessionFactory {
             } catch (final Exception ignored) {
                 // do nothing
             }
-
         }
     }
-
 }